Tutorial 3 home work, Accordion
