--- src/Tutorial-b2l2 copy/App.js
@@ -1,12 +1,0 @@
-import './App.css';
-import Form from "./components/Form";
-
-
-function App() {
-
-  return (
-    <Form />
-  );
-}
-
-export default App;
