+++ src/Tutorial-b2l2/App.css
@@ -1,0 +1,44 @@
+@import url('https://fonts.googleapis.com/css2?family=Inter:wght@400;600;800&display=swap');
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  background-color: #e4edfa;
+  font-family: 'Inter', sans-serif;
+  font-weight: 400;
+  font-size: 14px;
+  color: #000;
+}
+
+#app {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+}
+
+.app-container {
+  max-width: 550px;
+  padding: 53px 50px;
+  background: #ffffff;
+  border: 1px solid #d6e1f1;
+  box-shadow: 0px 15px 30px rgba(0, 0, 0, 0.04);
+  border-radius: 30px;
+  margin: 200px auto 0;
+}
+
+.app-title {
+  padding-left: 62px;
+  margin-bottom: 34px;
+  font-weight: 800;
+  font-size: 32px;
+  line-height: 39px;
+  /*background-image: url('./img/title-bg.svg');*/
+  background-repeat: no-repeat;
+  background-size: 54px;
+  background-position: left center;
+}
