+++ src/Tutorial-b2l2/components/Accordion/index.css
@@ -1,0 +1,72 @@
+
+/* Acordeon styles */
+.tab {
+  position: relative;
+  margin-bottom: 18px;
+  padding-bottom: 18px;
+  width: 100%;
+  overflow: hidden;
+  border-bottom: 1px solid #f3f3f3;
+}
+
+.tab:last-child {
+  border-bottom: 0;
+}
+
+.tab input {
+  position: absolute;
+  opacity: 0;
+  z-index: -1;
+}
+.tab label {
+  position: relative;
+  display: block;
+  background: #fff;
+  cursor: pointer;
+  font-weight: 600;
+  font-size: 18px;
+  line-height: 22px;
+  user-select: none;
+}
+.blue label {
+  background: #2980b9;
+}
+.tab-content {
+  max-height: 0;
+  overflow: hidden;
+  -webkit-transition: max-height 0.35s;
+  -o-transition: max-height 0.35s;
+  transition: max-height 0.35s;
+}
+.tab-content p {
+  color: #6d6d6d;
+  padding: 14px 0 24px;
+  line-height: 22px;
+}
+/* :active */
+.tab.active .tab-content {
+  max-height: 100vh;
+}
+/* Icon */
+.tab label::after {
+  position: absolute;
+  content: '';
+  width: 16px;
+  height: 10px;
+  background-image: url('./img/arrow-tab.svg');
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: 16px;
+  right: 0;
+  top: 0px;
+  display: block;
+  line-height: 3;
+  text-align: center;
+  -webkit-transition: all 0.35s;
+  -o-transition: all 0.35s;
+  transition: all 0.35s;
+}
+
+.tab.active label::after {
+  transform: rotate(180deg);
+}
