--- src/Tutorial-b2l2/components/Form/index.css
@@ -1,12 +1,0 @@
-.form-wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: column;
-  height: 100vh;
-}
-
-.form-wrapper input {
-  margin-bottom: 20px;
-}
